feat(commitlint): install commitlint
Commitlint lints the actual commit messages to ensure they are in
the required format. This allows developers that are experienced
with the required format to make commits without the use of the
Comitizen prompt - saving time and still ensuring the commit
format requirements are met.

It is important to have both Commitizen and Commitlint operating -
commits made with Commitizen can still be editted after the prompt
completes!